📝 (README.md): update GitHub profile visuals and organization
Enhance the visual appeal and organization of the README by updating
the theme of GitHub stats and trophies to a consistent dark theme.
Remove unnecessary comments for a cleaner look. Improve the layout
by removing redundant sections and optimizing the display of
GitHub activity. These changes aim to create a more professional
and cohesive presentation of the GitHub profile, making it more
engaging and easier to navigate for visitors.